Moved context classes to `Context` namespace
