+++ test/functional/Context/AbstractContextTest.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Dhii\Expression\Test;
+namespace Dhii\Expression\Test\Context;
 
+use Dhii\Expression\Context\AbstractContext;
 use Xpmock\TestCase;
 
 /**
- * Tests {@see Dhii\Expression\AbstractContext}.
+ * Tests {@see Dhii\Expression\Context\AbstractContext}.
  *
  * @since 0.1
  */
@@ -16,14 +17,14 @@ class AbstractContextTest extends TestCase
      *
      * @since 0.1
      */
-    const TEST_SUBJECT_CLASSNAME = '\\Dhii\\Expression\\AbstractContext';
+    const TEST_SUBJECT_CLASSNAME = '\\Dhii\\Expression\\Context\\AbstractContext';
 
     /**
      * Creates a new instance of the test subject.
      *
      * @since [*next=version*]
      *
-     * @return Dhii\Expression\AbstractContext
+     * @return AbstractContext
      */
     public function createInstance()
     {
